[#2079] moved configuration to test only
