+++ Realization/backend/idm-extras/src/main/java/eu/bcvsolutions/idm/extras/event/processor/tree/NewTreeNodeProcessor.java
@@ -7,30 +7,21 @@ import org.springframework.context.annotation.Description;
 import org.springframework.stereotype.Component;
 import org.springframework.util.Assert;
 
-import com.google.common.collect.ImmutableMap;
-
 import eu.bcvsolutions.idm.core.api.dto.IdmIdentityDto;
 import eu.bcvsolutions.idm.core.api.dto.IdmTreeNodeDto;
 import eu.bcvsolutions.idm.core.api.event.AbstractEntityEventProcessor;
 import eu.bcvsolutions.idm.core.api.event.DefaultEventResult;
 import eu.bcvsolutions.idm.core.api.event.EntityEvent;
 import eu.bcvsolutions.idm.core.api.event.EventResult;
-import eu.bcvsolutions.idm.core.api.exception.ResultCodeException;
 import eu.bcvsolutions.idm.core.api.service.ConfigurationService;
 import eu.bcvsolutions.idm.core.api.service.IdmConfigurationService;
 import eu.bcvsolutions.idm.core.api.service.IdmIdentityService;
 import eu.bcvsolutions.idm.core.model.event.TreeNodeEvent;
 import eu.bcvsolutions.idm.core.notification.api.domain.NotificationLevel;
 import eu.bcvsolutions.idm.core.notification.api.dto.IdmMessageDto;
-import eu.bcvsolutions.idm.core.notification.api.dto.IdmNotificationTemplateDto;
-import eu.bcvsolutions.idm.core.notification.api.dto.NotificationConfigurationDto;
-import eu.bcvsolutions.idm.core.notification.api.service.IdmNotificationConfigurationService;
-import eu.bcvsolutions.idm.core.notification.api.service.IdmNotificationTemplateService;
 import eu.bcvsolutions.idm.core.notification.api.service.NotificationManager;
-import eu.bcvsolutions.idm.core.notification.entity.IdmEmailLog;
 import eu.bcvsolutions.idm.extras.ExtrasModuleDescriptor;
 import eu.bcvsolutions.idm.extras.config.domain.ExtrasConfiguration;
-import eu.bcvsolutions.idm.extras.domain.ExtrasResultCode;
 
 /**
  * After creation of new tree node sends notification to authority
@@ -53,10 +44,6 @@ public class NewTreeNodeProcessor extends AbstractEntityEventProcessor<IdmTreeNo
 	ExtrasConfiguration extrasConfiguration;
 	@Autowired
 	private ConfigurationService configurationService;
-	@Autowired
-	private IdmNotificationConfigurationService notificationConfigurationService;
-	@Autowired
-	private IdmNotificationTemplateService notificationTemplateService;
 
 	@Autowired
 	public NewTreeNodeProcessor(NotificationManager notificationManager,
@@ -81,18 +68,12 @@ public class NewTreeNodeProcessor extends AbstractEntityEventProcessor<IdmTreeNo
 		IdmTreeNodeDto treeNode = event.getContent();
 		//
 		String roleName = configurationService.getValue(TREE_NODE_CREATE_ROLE);
+		if (roleName == null) {
+			return new DefaultEventResult<>(event, this);
+		}
 		List<IdmIdentityDto> recipients = identityService.findAllByRoleName(roleName);
 		//
-		if (treeNode != null && !recipients.isEmpty()) {
-			NotificationConfigurationDto config =
-					notificationConfigurationService
-							.getConfigurationByTopicLevelNotificationType(
-									ExtrasModuleDescriptor.TOPIC_NEW_TREE_NODE,
-									NotificationLevel.SUCCESS,
-									IdmEmailLog.NOTIFICATION_TYPE);
-			if (config == null) {
-				createNotificationConfiguration();
-			}
+		if (treeNode != null) {
 			notificationManager.send(
 					ExtrasModuleDescriptor.TOPIC_NEW_TREE_NODE,
 					new IdmMessageDto.Builder()
@@ -111,22 +92,5 @@ public class NewTreeNodeProcessor extends AbstractEntityEventProcessor<IdmTreeNo
 	public int getOrder() {
 		// after create
 		return 1000;
-	}
-
-	private void createNotificationConfiguration() {
-		NotificationConfigurationDto notificationConfiguration = new NotificationConfigurationDto();
-		notificationConfiguration.setTopic(ExtrasModuleDescriptor.TOPIC_NEW_TREE_NODE);
-		notificationConfiguration.setNotificationType(IdmEmailLog.NOTIFICATION_TYPE);
-		notificationConfiguration.setLevel(NotificationLevel.SUCCESS);
-
-		//get notification template
-		IdmNotificationTemplateDto idmNotificationTemplateDto =
-				notificationTemplateService.getByCode(ExtrasModuleDescriptor.TOPIC_NEW_TREE_NODE);
-		if (idmNotificationTemplateDto == null) {
-			throw new ResultCodeException(ExtrasResultCode.TEMPLATE_NOT_FOUND,
-					ImmutableMap.of("code", ExtrasModuleDescriptor.TOPIC_NEW_TREE_NODE));
-		}
-		notificationConfiguration.setTemplate(idmNotificationTemplateDto.getId());
-		notificationConfigurationService.save(notificationConfiguration);
 	}
 }
