+++ Realization/backend/idm-extras/src/test/java/eu/bcvsolutions/idm/extras/event/processor/tree/NewTreeNodeProcessorTest.java
@@ -1,21 +1,26 @@
 package eu.bcvsolutions.idm.extras.event.processor.tree;
 
 import static org.junit.Assert.assertEquals;
-import static org.junit.Assert.assertNotNull;
 
 import org.junit.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 
+import com.google.common.collect.ImmutableMap;
+
 import eu.bcvsolutions.idm.core.api.dto.IdmIdentityDto;
 import eu.bcvsolutions.idm.core.api.dto.IdmRoleDto;
+import eu.bcvsolutions.idm.core.api.exception.ResultCodeException;
 import eu.bcvsolutions.idm.core.api.service.ConfigurationService;
+import eu.bcvsolutions.idm.core.notification.api.domain.NotificationLevel;
 import eu.bcvsolutions.idm.core.notification.api.dto.IdmNotificationTemplateDto;
+import eu.bcvsolutions.idm.core.notification.api.dto.NotificationConfigurationDto;
 import eu.bcvsolutions.idm.core.notification.api.dto.filter.IdmNotificationFilter;
 import eu.bcvsolutions.idm.core.notification.api.service.IdmNotificationConfigurationService;
 import eu.bcvsolutions.idm.core.notification.api.service.IdmNotificationLogService;
 import eu.bcvsolutions.idm.core.notification.api.service.IdmNotificationTemplateService;
 import eu.bcvsolutions.idm.core.notification.entity.IdmEmailLog;
 import eu.bcvsolutions.idm.extras.ExtrasModuleDescriptor;
+import eu.bcvsolutions.idm.extras.domain.ExtrasResultCode;
 import eu.bcvsolutions.idm.test.api.AbstractIntegrationTest;
 import eu.bcvsolutions.idm.test.api.TestHelper;
 
@@ -28,6 +33,8 @@ public class NewTreeNodeProcessorTest extends AbstractIntegrationTest {
 	@Autowired
 	private IdmNotificationLogService notificationLogService;
 	@Autowired
+	private IdmNotificationConfigurationService notificationConfigurationService;
+	@Autowired
 	private IdmNotificationTemplateService notificationTemplateService;
 
 
@@ -37,11 +44,11 @@ public class NewTreeNodeProcessorTest extends AbstractIntegrationTest {
 
 		configurationService.setValue(NewTreeNodeProcessor.TREE_NODE_CREATE_ROLE, roleName);
 
-		// TODO set config for template
-
 		IdmRoleDto role = testHelper.createRole(roleName);
 		IdmIdentityDto identityDto = testHelper.createIdentity("TestIdentityNotify");
 		testHelper.createIdentityRole(identityDto, role);
+
+		createNotificationConfiguration();
 
 		//test send
 		testHelper.createTreeNode();
@@ -56,36 +63,48 @@ public class NewTreeNodeProcessorTest extends AbstractIntegrationTest {
 	}
 
 	@Test
-	public void createTestAndException() {
-		String roleName = "TestRoleNameNotifyExc";
-
-		configurationService.setValue(NewTreeNodeProcessor.TREE_NODE_CREATE_ROLE, roleName);
-
-		// TODO set config for template
+	public void createTestAndNotSendNotification() {
+		String roleName = "TestRoleNameNotify002";
 
 		IdmRoleDto role = testHelper.createRole(roleName);
-		IdmIdentityDto identityDto = testHelper.createIdentity("TestIdentityNotifyExc");
+		IdmIdentityDto identityDto = testHelper.createIdentity("TestIdentityNotify002");
 		testHelper.createIdentityRole(identityDto, role);
 
-		IdmNotificationTemplateDto template =
-				notificationTemplateService.getByCode(ExtrasModuleDescriptor.TOPIC_NEW_TREE_NODE);
-
-		assertNotNull(template);
-		String code = template.getCode();
-		template.setCode("TestCode0001");
-		notificationTemplateService.save(template);
+		createNotificationConfiguration();
 
 		//test send
-		Exception ex = null;
-		try {
-			testHelper.createTreeNode();
-		} catch (Exception e){
-			ex = e;
-		}
-		assertNotNull(ex);
+		testHelper.createTreeNode();
 
-		template.setCode(code);
-		notificationTemplateService.save(template);
+		IdmNotificationFilter filter = new IdmNotificationFilter();
+		filter.setRecipient(identityDto.getUsername());
+		filter.setNotificationType(IdmEmailLog.class);
+
+		long count = notificationLogService.count(filter);
+
+		assertEquals(0, count);
+	}
+
+	private void createNotificationConfiguration() {
+		NotificationConfigurationDto notificationConfiguration =
+				notificationConfigurationService.getConfigurationByTopicLevelNotificationType(
+						ExtrasModuleDescriptor.TOPIC_NEW_TREE_NODE,
+						NotificationLevel.SUCCESS,
+						IdmEmailLog.NOTIFICATION_TYPE);
+		if (notificationConfiguration == null) {
+			notificationConfiguration = new NotificationConfigurationDto();
+			notificationConfiguration.setTopic(ExtrasModuleDescriptor.TOPIC_NEW_TREE_NODE);
+			notificationConfiguration.setNotificationType(IdmEmailLog.NOTIFICATION_TYPE);
+			notificationConfiguration.setLevel(NotificationLevel.SUCCESS);
+
+			//get notification template
+			IdmNotificationTemplateDto idmNotificationTemplateDto =
+					notificationTemplateService.getByCode(ExtrasModuleDescriptor.TOPIC_NEW_TREE_NODE);
+			if (idmNotificationTemplateDto == null) {
+				throw new IllegalArgumentException("No template found!");
+			}
+			notificationConfiguration.setTemplate(idmNotificationTemplateDto.getId());
+			notificationConfigurationService.save(notificationConfiguration);
+		}
 	}
 
 }
