+++ Realization/frontend/czechidm-extras/src/locales/cs.json
@@ -47,10 +47,6 @@
     "COLUMN_NOT_FOUND": {
       "title": "Sloupec nenalezen.",
       "message": "Sloupec [{{name}}] nebyl nenalezen."
-    },
-    "TEMPLATE_NOT_FOUND": {
-      "title": "Šablona nalezena!",
-      "message": "Šablona s kódem [{{code}}] nebyla nalezena."
     }
   }
 }
